requirements.txt
streamlit
python-dotenv
langchain
langchain_community
openai
faiss-cpu
pandas
tiktoken
sentence-transformers
accelerate
multiprocess